Add user null check for remove method and Update Test(s)

+++ src/main/java/com/example/investmentportfoliorebalancingtool/domain/repositories/UserRepositoryImpl.java
@@ -66,6 +66,8 @@ public class UserRepositoryImpl implements UserRepository {
 
     @Override
     public void remove(User user) {
-        userDAO.delete(user);
+        if (user != null) {
+            userDAO.delete(user);
+        }
     }
 }
